refactor(gui): split fact partition/tests and resolve gate collisions
The D6/density/fact-filter fix grew territoryPartition.ts (610) and
territory-layout.vitest.ts (761) past the file-complexity caps. Extract the
fact partition into territoryPartitionFact.ts (shared daysAgo/types stay
authoritative in territoryPartition.ts, one-way import, no cycle) and move the
fact-skel suite into territory-layout-fact.vitest.ts with a shared
territory-layout-fixtures.ts; register the new suite in the GUI test manifest.
Also rename ledgerGraphLayout's private labelOf to entityLabel to clear a
duplicate-definition collision with canvasEgoLayout's labelOf.
