+++ tools/gui-test-manifest.mjs
@@ -11,7 +11,9 @@ export const guiVitestManifest = [
   "packages/gui/test/graphNavigation.vitest.ts",
   "packages/gui/test/canvas-ego-layout.vitest.ts",
   "packages/gui/test/territory-layout.vitest.ts",
+  "packages/gui/test/territory-layout-fact.vitest.ts",
   "packages/gui/test/ledger-graph-layout.vitest.ts",
+  "packages/gui/test/d6-focus-continuity.vitest.ts",
   "packages/gui/test/entitySearch.vitest.ts",
   "packages/gui/test/commandPalette-render.vitest.ts"
 ];
